+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 ﻿# Changelog
 
+## [1.1.1] - 2024-04-02
+### 🚀 New Features
+No features
+
+### 🐛 Bug Fixes
+- All tooltips in the cards have been translated
+- All cards now have a customizable title
+
+### ⚠️ Breaking Changes
+No breaking changes
+
+
 ## [1.1.0] - 2024-03-30
 ### 🚀 New Features
 - Icon added